Extension: Add CORS proxy and URL override support
This is useful for people who want to use the same URL as the server
URL in SillyTavern. All requests uses SillyTavern's CORS proxy setting
rather than requesting tabbyAPI from the client side.

An example, the user has http://localhost:5000 set as the URL for
TabbyAPI. However, the extension can't connect to localhost since
the client is a different device, with a different localhost. The
CORS proxy takes the requests and uses the server's "localhost" which
is connected to TabbyAPI.

Signed-off-by: kingbri <bdashore3@proton.me>